Instructors and Operators
Three types of instructions:

1. Type declaration instruction
2. Arithmetic instruction 
3. Control instruction

OUTPUT FOR THE TYPE DECLARATION INSTRUCTION PROGRAM:

The value of a is 4
The value of b is 4
The value of c is 4

Arithmetic instruction program:

In the following line (3,2) are operands and (*,+) are operators. float, int can be operands.
(3 * 2) + 2

% is a modular division operator which gives remainder as the output. For example 5%2 = 1.

Type Conversion In Arithmetic Instruction:

Outputs for the Arithmetic Conversion Example Program:

Operation between int and int results in int - for example 6 + 5 = 11 and 5/2 = 2 and 2/5 = 0 - %d is inputed in this case
Operation between int and float results in float - for example 6 + 5.6 = 11.600000 and 5/2 = 2.5 and 2/5 = 0.4 - %f is inputed in this case
Operation between float and float results in float - for example 6.1 + 5.6 = 11.700000

int a = 3.5; // 3.5(float) will be converted to 3(int) because a is not able to store float and therefore a will be stored as 3. So float is demoted to int
float a = 8;  // 8(int) will be converted to 8.0(float) because a is not able to store int and therefore a will be stored as 8.0. So int is promoted to float